增加確認檔案是否存在, multiprocessing 增加p.join()

+++ download_beauty.py
@@ -21,10 +21,13 @@ img_regex = re.compile(r'<img src="[^"]*(//[i|m].imgur.com/[^"]+jpg)" alt=')
 
 
 def download_pic(pic_url, dir):
-    # 尚未處理重複下載的問題
+    # 增加處理重複下載的問題
     try:
         pic_name = dir + '/' + pic_url.split('/')[-1]
-        urllib.urlretrieve(pic_url, pic_name)
+        if os.path.exists(os.getcwd() + '/' + pic_name): 
+            print pic_name + ",已經下載過了喔 +_+"
+        else:
+            urllib.urlretrieve(pic_url, pic_name)
     except IOError as ioerr:
         print "IOError in download picture: " + pic_url
 
@@ -83,11 +86,13 @@ def store_pic(url, rate=""):
 
     # 如何優化計數動作?
     # 預計管理 Process 數目
+    # 增加p.join()
     count = 0
     for pic_url in pic_url_list:
         count += 1
         p = multiprocessing.Process(target=download_pic, args=(pic_url, dir_name,))
         p.start()
+        p.join()
         # download_pic(pic_url, title)
 
     if count == 0:
